Move reflection related logging to own module.
Add KotlinLogging Facade.
